feat: crud operations for the locations without authorization
