+++ apps/api/src/location/location.module.ts
@@ -1,8 +1,10 @@
 import { Module } from '@nestjs/common';
 import { LocationService } from './location.service';
 import { LocationController } from './location.controller';
+import { DbModule } from 'src/db/db.module';
 
 @Module({
+  imports: [DbModule],
   controllers: [LocationController],
   providers: [LocationService],
 })
